Disabled ZIM Scan, Fixed Hash Sends, Improved Search Indexing
Disabled ZIM, leaving it as I still want to get it working later.
Updated HASH handling for seasons folders.
Search indexing only starts when the search bar is clicked as it can crash the device under the right cercumstances. will move this into the index handler in the backend eventualy.